Add feature flag system
Dynamically generate Feature.h which defines the enabled features. The
version info includes the list of compiled in flags.  A helper function
allows for checking if features are enabled.  (This is done as a
constexpr method so the checking will happen at compile time)

This makes for a more fine grained ability to turn on various features
at compile time.